Move contents of asset.rs into agent.rs
