+++ src/agent.rs
@@ -1,6 +1,6 @@
 #![allow(missing_docs)]
-use crate::asset::Asset;
 use crate::input::deserialise_proportion_nonzero;
+use crate::process::Process;
 use crate::region::RegionSelection;
 use anyhow::Result;
 use serde::Deserialize;
@@ -93,4 +93,17 @@ pub enum ObjectiveType {
     LevelisedCostOfX,
     #[string = "eac"]
     EquivalentAnnualCost,
+}
+
+/// An asset controlled by an agent.
+#[derive(Clone, Debug, PartialEq)]
+pub struct Asset {
+    /// The [Process] that this asset corresponds to
+    pub process: Rc<Process>,
+    /// The region in which the asset is located
+    pub region_id: Rc<str>,
+    /// Capacity of asset
+    pub capacity: f64,
+    /// The year the asset comes online
+    pub commission_year: u32,
 }
