+++ src/lib.rs
@@ -1,7 +1,6 @@
 //! High level functionality for launching the simulation.
 #![warn(missing_docs)]
 pub mod agent;
-pub mod asset;
 pub mod commodity;
 pub mod demand;
 pub mod input;
